DZ_12/internal/consumer/requirements.txt
pyspark==3.3.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
